use separate file for _PROXY_track_cost_callback
